Consume proposal flow flags
