--- components/[pageId]/DocumentPage/components/ProposalProperties/index.ts
@@ -1,3 +1,0 @@
-import ProposalProperties from './ProposalProperties';
-
-export { ProposalProperties };
