+++ components/proposals/components/ProposalStepper/index.ts
@@ -1,0 +1,1 @@
+export { ProposalStepper } from './ProposalStepper';
